requirements.txt
requests
pillow
tk
configparser
